Move macaroon tests from api to apiserver
